requirements.txt
accelerate
decord
einops
flash-attn==2.7.4.post1
numpy==1.26.4
open-clip-torch>=2.23.0
openai>=1.29.0
openpyxl
pandas>=2.1.4
Pillow
qwen_vl_utils
sentencepiece>=0.1.99
spacy==3.8.7
tiktoken
tqdm
transformers==4.49.0

git+https://github.com/openai/CLIP.git#egg=clip
git+https://github.com/LLaVA-VL/LLaVA-NeXT.git#egg=llava
